remove dependency of LanguageFunctions
